[FIX] .env 파일 추가, 커뮤니티 코드 정리
