+++ .env
@@ -1,6 +1,8 @@
-VITE_REACT_APP_SERVER_URL=https://www.purupuru.store
+VITE_REACT_APP_SERVER_URL=https://www.purupuru.store/
 VITE_REACT_APP_GOOGLE_ANALYTICS_TRACKING_ID="G-Y2HVS7W232"
-VITE_KAKAO_REDIRECT_URI=api/auth/login/kakao/return
-VITE_GOOGLE_REDIRECT_URI=api/auth/login/google/return
-VITE_GOOGLE_REST_API_ID=214149105868-8h686c0pdnk0cvscof9qr24604t5rdh8.apps.googleusercontent.com
+
+VITE_KAKAO_REDIRECT_URI=https://puru-puru.vercel.app/api/auth/login/kakao/return
 VITE_KAKAO_REST_API_KEY=19449fa93915b6797984debf2dada7a3
+
+VITE_GOOGLE_REDIRECT_URI=https://puru-puru.vercel.app/api/auth/login/google/return
+VITE_GOOGLE_OAUTH_CLIENT_ID=214149105868-8h686c0pdnk0cvscof9qr24604t5rdh8.apps.googleusercontent.com
